% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descarga-lee-Datos.R
\name{descargar_leer_datos_estacion}
\alias{descargar_leer_datos_estacion}
\title{Descarga y lee datos de estaciones}
\usage{
descargar_leer_datos_estacion(id_estacion, ruta_archivo)
}
\arguments{
\item{id_estacion}{Un identificador de estación meteorologica específica. Debe ser uno de los siguientes: "NH0472", "NH0910", "NH0046", "NH0098", "NH0437".}

\item{ruta_archivo}{Ruta de destino donde se guardara el archivo descargado.}
}
\value{
Un data frame que contiene los datos de estación meteorlógica seleccionada.
}
\description{
Esta función descarga datos de una estación meteorlóguca especifica y los lee en un data frame.
}
\examples{
descargar_leer_datos_estacion("NH0472", "data/NH0472.csv")


}
