% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_temperatura_mensual.R
\name{grafico_temperatura_mensual}
\alias{grafico_temperatura_mensual}
\title{Genera un gráfico de temperatura promedio mensual}
\usage{
grafico_temperatura_mensual(datos, colores = NULL, titulo = "Temperatura")
}
\arguments{
\item{datos}{Un data frame que contiene datos meteorlógicos. Debe incluir las columnas 'fecha', 'id' y 'temperatura_abrigo_150cm'.}

\item{colores}{Argumento para indicar que colores usamos oara el grafico.}

\item{titulo}{Titulo del grafico. El valor predeterminado es "Temperatura".}
}
\value{
Grafico que muestra el promedio mensual de la temperatura de abrigo.
}
\description{
Genera un gráfico de temperatura promedio mensual
}
\examples{
datos <- descargar_leer_datos_estacion("NH0472", "data/NH0472.csv")
grafico_temperatura_mensual(datos, titulo = "Temperatura Promedio Mensual")
print(grafico)


}
