% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_resumen_temperatura.R
\name{tabla_resumen_temperatura}
\alias{tabla_resumen_temperatura}
\title{Generar tabla resumen de temperatura para una o más estaciones.}
\usage{
tabla_resumen_temperatura(datos)
}
\arguments{
\item{datos}{Un data frame que contiene datos meteológicos. Dene incluir las columnas 'id' y 'temperatura_abrigo_150cm'.}
}
\value{
Un dataframe con el resumen de temperatura para cada estación, que incluye columnas para la temperatura mínima (\code{min_temp}), máxima (\code{max_temp}), y promedio (\code{mean_temp}).
}
\description{
Esta función crea una tabla resumen de la temperatura registrada en la columna 'temperatura_abrigo_150cm' para una o más estaciones, mostrando los valores mínimos, máximos y el promedio.
}
\examples{
datos <- descargar_leer_datos_estacion("NH0472", "data/NH0472.csv")
tabla_resumen_temperatura(datos)


}
