% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descarga-lee-Datos.R
\name{descargar_leer_datos_estacion}
\alias{descargar_leer_datos_estacion}
\title{Descarga y lee datos de estaciones}
\usage{
descargar_leer_datos_estacion(id_estacion)
}
\arguments{
\item{id_estacion}{Un identificador de estación meteorologica específica. Debe ser uno de los siguientes: "NH0472", "NH0910", "NH0046", "NH0098", "NH0437".}
}
\value{
Un data frame que contiene los datos de estación meteorlógica seleccionada.
}
\description{
Esta función descarga datos de una estación meteorlógica especifica y los lee en un data frame.
}
\examples{
descargar_leer_datos_estacion("NH0046")


}
